move fixture recipes (makes it easier for devs to write new ones)
