+++ app/services/standard_edition_migrator/fixtures/legacy_presenter.rb
@@ -1,0 +1,16 @@
+class StandardEditionMigrator::LegacyPresenter
+  def initialize(_legacy_record, update_type: nil, title: nil); end
+
+  def content
+    {
+      "body": "OLD PAYLOAD",
+      "some_old_field": "some old value",
+    }
+  end
+
+  def links
+    {
+      "old_link": "old link",
+    }
+  end
+end
