+++ test/unit/app/services/standard_edition_migrator_test.rb
@@ -1,4 +1,7 @@
 require "test_helper"
+require_relative "../../../../app/services/standard_edition_migrator/fixtures/legacy_presenter"
+require_relative "../../../../app/services/standard_edition_migrator/fixtures/recipe_for_legacy_editionable_document"
+require_relative "../../../../app/services/standard_edition_migrator/fixtures/recipe_for_non_editionable_record"
 
 class StandardEditionMigratorTest < ActiveSupport::TestCase
   extend Minitest::Spec::DSL
@@ -17,7 +20,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
     test "raises exception if an Edition is passed" do
       edition = build(:edition)
       assert_raises(RuntimeError, "An Edition was passed. You must pass the Document instead (so that we can migrate all of its Editions)") do
-        StandardEditionMigrator.preview_migration(edition, RecipeForLegacyEditionableDocument)
+        StandardEditionMigrator.preview_migration(edition, StandardEditionMigrator::RecipeForLegacyEditionableDocument)
       end
     end
 
@@ -80,17 +83,17 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
       OUTPUT
 
       # Test with a legacy record which is Editionable (i.e. Document) - should preview migration of the latest Edition
-      summary = StandardEditionMigrator.preview_migration(@legacy_editionable_document, RecipeForLegacyEditionableDocument)
+      summary = StandardEditionMigrator.preview_migration(@legacy_editionable_document, StandardEditionMigrator::RecipeForLegacyEditionableDocument)
       assert_equal expected_output, summary.chomp
 
       # Test with a legacy record which is not Editionable - should preview migration of the record itself
-      summary = StandardEditionMigrator.preview_migration(@legacy_non_editionable_record, RecipeForNonEditionableRecord)
+      summary = StandardEditionMigrator.preview_migration(@legacy_non_editionable_record, StandardEditionMigrator::RecipeForNonEditionableRecord)
       assert_equal expected_output, summary.chomp
     end
 
     test "doesn't create any StandardEdition or Document, and doesn't persist any changes to the legacy record" do
       assert_no_difference [StandardEdition.method(:count), Document.method(:count)] do
-        StandardEditionMigrator.preview_migration(@legacy_editionable_document, RecipeForLegacyEditionableDocument)
+        StandardEditionMigrator.preview_migration(@legacy_editionable_document, StandardEditionMigrator::RecipeForLegacyEditionableDocument)
       end
 
       assert_equal "Detailed Guide", @legacy_editionable_document.reload.editions.last.title
@@ -99,7 +102,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
 
   describe "#create_new_document" do
     test "performs the migration and saves the new edition on a legacy non-editionable record" do
-      StandardEditionMigrator.create_new_document(@legacy_non_editionable_record, RecipeForNonEditionableRecord)
+      StandardEditionMigrator.create_new_document(@legacy_non_editionable_record, StandardEditionMigrator::RecipeForNonEditionableRecord)
       edition = StandardEdition.last
       assert edition.persisted?
       assert_equal "test_type", edition.configurable_document_type
@@ -110,7 +113,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
     test "raises exception if a Document is passed (we could handle this in theory, but for simplicity we expect all Document conversions to go through the migrate_existing_document route)" do
       document = build(:document)
       error = assert_raises(RuntimeError) do
-        StandardEditionMigrator.create_new_document(document, RecipeForLegacyEditionableDocument)
+        StandardEditionMigrator.create_new_document(document, StandardEditionMigrator::RecipeForLegacyEditionableDocument)
       end
 
       assert_equal "Cannot pass a Document to create_new_document", error.message
@@ -120,7 +123,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
       error = nil
       assert_no_difference [StandardEdition.method(:count), Document.method(:count)] do
         error = assert_raises(RuntimeError) do
-          StandardEditionMigrator.create_new_document(@legacy_non_editionable_record, RecipeForNonEditionableRecord, raise_if_payloads_differ: true)
+          StandardEditionMigrator.create_new_document(@legacy_non_editionable_record, StandardEditionMigrator::RecipeForNonEditionableRecord, raise_if_payloads_differ: true)
         end
       end
 
@@ -132,7 +135,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
     test "performs the migration and saves the new edition (and document) on a legacy editionable document" do
       old_id = @legacy_editionable_document.content_id
       old_body = @legacy_editionable_document.editions.last.body
-      StandardEditionMigrator.migrate_existing_document(@legacy_editionable_document, RecipeForLegacyEditionableDocument)
+      StandardEditionMigrator.migrate_existing_document(@legacy_editionable_document, StandardEditionMigrator::RecipeForLegacyEditionableDocument)
       edition = StandardEdition.last
       assert edition.persisted?
       assert_equal "test_type", edition.configurable_document_type
@@ -144,7 +147,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
     test "raises exception if a non-Document is passed" do
       non_document = build(:organisation)
       error = assert_raises(RuntimeError) do
-        StandardEditionMigrator.migrate_existing_document(non_document, RecipeForLegacyEditionableDocument)
+        StandardEditionMigrator.migrate_existing_document(non_document, StandardEditionMigrator::RecipeForLegacyEditionableDocument)
       end
 
       assert_equal "Cannot pass a non-Document to migrate_existing_document", error.message
@@ -154,7 +157,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
       error = nil
       assert_no_difference [StandardEdition.method(:count), Document.method(:count)] do
         error = assert_raises(RuntimeError) do
-          StandardEditionMigrator.migrate_existing_document(@legacy_editionable_document, RecipeForLegacyEditionableDocument, raise_if_payloads_differ: true)
+          StandardEditionMigrator.migrate_existing_document(@legacy_editionable_document, StandardEditionMigrator::RecipeForLegacyEditionableDocument, raise_if_payloads_differ: true)
         end
       end
 
@@ -171,7 +174,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
       ]
       StandardEditionMigrator.enqueue_bulk_migration(
         legacy_non_editionable_records,
-        RecipeForNonEditionableRecord,
+        StandardEditionMigrator::RecipeForNonEditionableRecord,
         migration_method: "create_new_document",
       )
 
@@ -182,7 +185,7 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
       keyword_args = first_job_args.second
       assert_equal legacy_non_editionable_records.first.id, record_id
       assert_equal "Organisation", keyword_args["model_class"]
-      assert_equal "StandardEditionMigratorTest::RecipeForNonEditionableRecord", keyword_args["recipe_class"]
+      assert_equal "StandardEditionMigrator::RecipeForNonEditionableRecord", keyword_args["recipe_class"]
       assert_equal "create_new_document", keyword_args["migration_method"]
 
       second_job_args = StandardEditionMigratorJob.jobs.second["args"]
@@ -190,106 +193,8 @@ class StandardEditionMigratorTest < ActiveSupport::TestCase
       keyword_args = second_job_args.second
       assert_equal legacy_non_editionable_records.second.id, record_id
       assert_equal "Organisation", keyword_args["model_class"]
-      assert_equal "StandardEditionMigratorTest::RecipeForNonEditionableRecord", keyword_args["recipe_class"]
+      assert_equal "StandardEditionMigrator::RecipeForNonEditionableRecord", keyword_args["recipe_class"]
       assert_equal "create_new_document", keyword_args["migration_method"]
-    end
-  end
-
-  class RecipeForLegacyEditionableDocument < StandardEditionMigrator::BaseRecipe
-    def legacy_presenter
-      LegacyPresenter
-    end
-
-    def build_edition(legacy_record)
-      edition_attrs = {
-        configurable_document_type: "test_type",
-        updated_at: legacy_record.updated_at.rfc3339,
-        first_published_at: legacy_record.created_at.rfc3339,
-        major_change_published_at: legacy_record.updated_at.rfc3339,
-        creator: User.last,
-        document: legacy_record.document,
-      }
-      @edition = StandardEdition.new(edition_attrs)
-
-      legacy_record.translations.each do |translation|
-        # Still operating on the newly initialized Edition in memory - careful use of `find_or_initialize_by`
-        @edition.translations.find_or_initialize_by(locale: translation.locale).update(
-          title: title(translation),
-          summary: summary(translation),
-          block_content: {
-            "field_attribute" => translation.body.to_s,
-          },
-        )
-      end
-
-      @artefacts_to_save = @edition.translations
-      @edition
-    end
-
-    def title(_legacy_record)
-      "Title"
-    end
-
-    def summary(_legacy_record)
-      "summary"
-    end
-  end
-
-  class RecipeForNonEditionableRecord < StandardEditionMigrator::BaseRecipe
-    def legacy_presenter
-      LegacyPresenter
-    end
-
-    def build_edition(legacy_record)
-      edition_attrs = {
-        configurable_document_type: "test_type",
-        updated_at: legacy_record.updated_at.rfc3339,
-        first_published_at: legacy_record.created_at.rfc3339,
-        major_change_published_at: legacy_record.updated_at.rfc3339,
-        creator: User.last,
-      }
-      @edition = StandardEdition.new(edition_attrs)
-
-      # NOTE: implementation will vary depending on non-editionable model.
-      # Organisation has `translations` we can iterate over. TopicalEvent does not.
-      legacy_record.translations.each do |translation|
-        @edition.translations.find_or_initialize_by(locale: translation.locale).update(
-          title: translation.name,
-          summary: summary(legacy_record),
-          block_content: {
-            # Hardcoded for simplicity, so we can check the payload comes out the same
-            # for both the editionable and non-editionable test cases.
-            "field_attribute" => "Old body",
-          },
-        )
-      end
-      @artefacts_to_save = @edition.translations
-      @edition
-    end
-
-    def title(_legacy_record)
-      "Title"
-    end
-
-    def summary(_legacy_record)
-      "summary"
-    end
-  end
-
-  class LegacyPresenter
-    def initialize(_legacy_record, update_type: nil, title: nil); end
-
-    def content
-      {
-        "body": "OLD PAYLOAD",
-        "some_old_field": "some old value",
-      }
-    end
-
-    def links
-      {
-        "old_link": "old link",
-      }
     end
   end
 end
